Refactor river network format handling into a class-based registry (#196)
* Refactor river network creation to class-based format registry

* chore: improve error message

* chore: please qa

---------
