+++ src/earthkit/hydro/river_network/_cache.py
@@ -4,8 +4,8 @@ from functools import wraps
 from hashlib import sha256
 
 import joblib
-from earthkit.hydro._version import __version__ as ekh_version
 
+from earthkit.hydro._version import __version__ as ekh_version
 from earthkit.hydro.data_structures._network import RiverNetwork
 
 # read in only up to second decimal point
